requirements.txt
asgiref==3.5.0
astroid==2.11.2
autopep8==1.6.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
dill==0.3.4
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-jazzmin==2.5.0
django-lifecycle==0.9.6
django-model-utils==4.2.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-split-settings==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
firebase-admin==5.2.0
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.3
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.46
Pillow==9.1.0
platformdirs==2.5.1
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.13.5
pyparsing==3.0.8
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.2.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
urlman==2.0.1
wrapt==1.14.0
